[AIRFLOW-6324] Hide celery commands when executor is not CeleryExecutor
