+++ airflow/cli/commands/celery_command.py
@@ -1,3 +1,4 @@
+#
 # Licensed to the Apache Software Foundation (ASF) under one
 # or more contributor license agreements.  See the NOTICE file
 # distributed with this work for additional information
@@ -14,7 +15,7 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-"""Worker command"""
+"""Celery command"""
 import os
 import signal
 import sys
@@ -22,13 +23,62 @@ from multiprocessing import Process
 from typing import Optional
 
 import daemon
+from celery.bin import worker as worker_bin
 from daemon.pidfile import TimeoutPIDLockFile
 
 from airflow import settings
 from airflow.configuration import conf
+from airflow.executors.celery_executor import app as celery_app
 from airflow.utils import cli as cli_utils
 from airflow.utils.cli import setup_locations, setup_logging, sigint_handler
 from airflow.utils.serve_logs import serve_logs
+
+
+@cli_utils.action_logging
+def flower(args):
+    """Starts Flower, Celery monitoring tool"""
+    broka = conf.get('celery', 'BROKER_URL')
+    address = '--address={}'.format(args.hostname)
+    port = '--port={}'.format(args.port)
+    api = ''  # pylint: disable=redefined-outer-name
+    if args.broker_api:
+        api = '--broker_api=' + args.broker_api
+
+    url_prefix = ''
+    if args.url_prefix:
+        url_prefix = '--url-prefix=' + args.url_prefix
+
+    basic_auth = ''
+    if args.basic_auth:
+        basic_auth = '--basic_auth=' + args.basic_auth
+
+    flower_conf = ''
+    if args.flower_conf:
+        flower_conf = '--conf=' + args.flower_conf
+
+    if args.daemon:
+        pid, stdout, stderr, _ = setup_locations("flower", args.pid, args.stdout, args.stderr, args.log_file)
+        stdout = open(stdout, 'w+')
+        stderr = open(stderr, 'w+')
+
+        ctx = daemon.DaemonContext(
+            pidfile=TimeoutPIDLockFile(pid, -1),
+            stdout=stdout,
+            stderr=stderr,
+        )
+
+        with ctx:
+            os.execvp("flower", ['flower', '-b',
+                                 broka, address, port, api, flower_conf, url_prefix, basic_auth])
+
+        stdout.close()
+        stderr.close()
+    else:
+        signal.signal(signal.SIGINT, sigint_handler)
+        signal.signal(signal.SIGTERM, sigint_handler)
+
+        os.execvp("flower", ['flower', '-b',
+                             broka, address, port, api, flower_conf, url_prefix, basic_auth])
 
 
 def _serve_logs(skip_serve_logs: bool = False) -> Optional[Process]:
@@ -50,17 +100,13 @@ def worker(args):
         print("Worker exiting... database connection precheck failed! ")
         sys.exit(1)
 
-    # Celery worker
-    from airflow.executors.celery_executor import app as celery_app
-    from celery.bin import worker  # pylint: disable=redefined-outer-name
-
     autoscale = args.autoscale
     skip_serve_logs = args.skip_serve_logs
 
     if autoscale is None and conf.has_option("celery", "worker_autoscale"):
         autoscale = conf.get("celery", "worker_autoscale")
 
-    worker = worker.worker(app=celery_app)   # pylint: disable=redefined-outer-name
+    worker_instance = worker_bin.worker(app=celery_app)
     options = {
         'optimization': 'fair',
         'O': 'fair',
@@ -92,7 +138,7 @@ def worker(args):
         )
         with ctx:
             sub_proc = _serve_logs(skip_serve_logs)
-            worker.run(**options)
+            worker_instance.run(**options)
 
         stdout.close()
         stderr.close()
@@ -101,7 +147,7 @@ def worker(args):
         signal.signal(signal.SIGTERM, sigint_handler)
 
         sub_proc = _serve_logs(skip_serve_logs)
-        worker.run(**options)
+        worker_instance.run(**options)
 
     if sub_proc:
         sub_proc.terminate()
