--- airflow/cli/commands/flower_command.py
@@ -1,75 +1,0 @@
-# -*- coding: utf-8 -*-
-#
-# Licensed to the Apache Software Foundation (ASF) under one
-# or more contributor license agreements.  See the NOTICE file
-# distributed with this work for additional information
-# regarding copyright ownership.  The ASF licenses this file
-# to you under the Apache License, Version 2.0 (the
-# "License"); you may not use this file except in compliance
-# with the License.  You may obtain a copy of the License at
-#
-#   http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing,
-# software distributed under the License is distributed on an
-# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-# KIND, either express or implied.  See the License for the
-# specific language governing permissions and limitations
-# under the License.
-"""Flower command"""
-import os
-import signal
-
-import daemon
-from daemon.pidfile import TimeoutPIDLockFile
-
-from airflow.configuration import conf
-from airflow.utils import cli as cli_utils
-from airflow.utils.cli import setup_locations, sigint_handler
-
-
-@cli_utils.action_logging
-def flower(args):
-    """Starts Flower, Celery monitoring tool"""
-    broka = conf.get('celery', 'BROKER_URL')
-    address = '--address={}'.format(args.hostname)
-    port = '--port={}'.format(args.port)
-    api = ''  # pylint: disable=redefined-outer-name
-    if args.broker_api:
-        api = '--broker_api=' + args.broker_api
-
-    url_prefix = ''
-    if args.url_prefix:
-        url_prefix = '--url-prefix=' + args.url_prefix
-
-    basic_auth = ''
-    if args.basic_auth:
-        basic_auth = '--basic_auth=' + args.basic_auth
-
-    flower_conf = ''
-    if args.flower_conf:
-        flower_conf = '--conf=' + args.flower_conf
-
-    if args.daemon:
-        pid, stdout, stderr, _ = setup_locations("flower", args.pid, args.stdout, args.stderr, args.log_file)
-        stdout = open(stdout, 'w+')
-        stderr = open(stderr, 'w+')
-
-        ctx = daemon.DaemonContext(
-            pidfile=TimeoutPIDLockFile(pid, -1),
-            stdout=stdout,
-            stderr=stderr,
-        )
-
-        with ctx:
-            os.execvp("flower", ['flower', '-b',
-                                 broka, address, port, api, flower_conf, url_prefix, basic_auth])
-
-        stdout.close()
-        stderr.close()
-    else:
-        signal.signal(signal.SIGINT, sigint_handler)
-        signal.signal(signal.SIGTERM, sigint_handler)
-
-        os.execvp("flower", ['flower', '-b',
-                             broka, address, port, api, flower_conf, url_prefix, basic_auth])
