+++ airflow/config_templates/default_airflow.cfg
@@ -399,13 +399,13 @@ sentry_dsn =
 celery_app_name = airflow.executors.celery_executor
 
 # The concurrency that will be used when starting workers with the
-# "airflow worker" command. This defines the number of task instances that
+# "airflow celery worker" command. This defines the number of task instances that
 # a worker will take, so size up your workers based on the resources on
 # your worker box and the nature of your tasks
 worker_concurrency = 16
 
 # The maximum and minimum concurrency that will be used when starting workers with the
-# "airflow worker" command (always keep minimum processes, but grow to maximum if necessary).
+# "airflow celery worker" command (always keep minimum processes, but grow to maximum if necessary).
 # Note the value should be "max_concurrency,min_concurrency"
 # Pick these numbers based on resources on worker box and the nature of the task.
 # If autoscale option is available, worker_concurrency will be ignored.
